refactor(api): move redis monitor to src/shared
