+++ api/tests/shared/unit/application/healthcheck/healthcheck-controller_test.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-restricted-paths
 import { knex } from '../../../../../db/knex-database-connection.js';
-import { redisMonitor } from '../../../../../lib/infrastructure/utils/redis-monitor.js';
 import { healthcheckController } from '../../../../../src/shared/application/healthcheck/healthcheck-controller.js';
+import { redisMonitor } from '../../../../../src/shared/infrastructure/utils/redis-monitor.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | healthcheckController', function () {
